pyproject.toml
[project]
name = "u-s-market-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.56.0",
    "jwt>=1.4.0",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]
